upd telegramId field type (int -> Long)

+++ bot/src/main/java/edu/java/bot/entities/User.java
@@ -7,6 +7,6 @@ import lombok.Getter;
 @Getter public class User {
     private String firstName;
     private String lastName;
-    private int telegramId;
+    private Long telegramId;
     private boolean isBot;
 }
